metadata_host is marked as required upstream but actually isn't (#1233)
Resolves #362 - users of this provider need to specify this config even
if the default is fine.

The upstream provider [has this property marked as required](https://github.com/hashicorp/terraform-provider-azuread/blob/6594e1c6cd59ffc7f5e9a881412609d9cde816e7/internal/provider/provider.go#L126),
but that works differently in Terraform: since there's also a
`DefaultFunc` configured, the function's return value will be used if
the user doesn't specify a value, and the function itself returns a
default of `""`. So a value is guaranteed, even if it's the empty
string.

In Pulumi, "required" means needs to be configured by the user.
Therefore, we mark this property as optional to get the same behavior as
upstream.